Fixes to match PCAP packets with SDP
We want to be able to match up the source and target <IP>:<PORT> with
the IP and port info sent in SDP. I adjusted the source and target
<IP>:<PORT> we write to the IP and UDP headers in the PCAP file.

I added the offer SDP (in addition to the already written answer SDP)
to the metadata file.

As a consequence of writing out the offer SDP, we will have empty PCAP
files and associated metadata files for calls that are never answered.